Move test code for Discriminator
 - Discriminator의 test code를 test_models.py 에서 test_modules.py로 이전
